Refactoring error handlers(auth)
